pyproject.toml
[tool.poetry]
name = "bot-telegram"
version = "0.1.0"
description = ""
authors = ["Rômulo do Carmo Sousa <carmosousa20@outlook.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = "^13.1"
python-dotenv = "^0.15.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
yapf = "^0.31.0"
mypy = "^0.790"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
